Corrected type for touch integer

+++ sel_compiler/src/tests_touch.rs
@@ -67,7 +67,7 @@ mod tests {
         let tree = compiler.compile(&input);
 
         let root = tree.get_root();
-        let root_value: i64 = tree.get_integer_value_of(&root).unwrap();
+        let root_value: i32 = tree.get_integer_value_of(&root).unwrap();
 
         assert_eq!(root.get_operation(), Operation::Touch);
         assert_eq!(root.get_data_type(), DataType::Integer);
